+++ CHANGELOG.md
@@ -6,10 +6,13 @@
 - XOT-922 Added Marketing Article Page template
 - CI-304 - Gave opportunity page a related region page and a scale numerical value to use for filtering
 - CI-267 - Added cta text and link for sector page for related opportunities section
+- CMS-1727 - Move region pages from Invest to International
 - CI-324 - Made AbstractSector page and sector and sub-sector page inherit from this and opportunity listing page filters by sub sector
 
 ### Fixed Bugs
 - NOTICKET - Capital invest landing page image serializers use aspect ratio for card images
+- CMS-1737 - Fix 500 on user approval
+
 
 ## [2019.06.27](https://github.com/uktrade/directory-cms/releases/tag/2019.06.27)
 [Full Changelog](https://github.com/uktrade/directory-cms/compare/2019.05.20...2019.06.27